% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{residuals.fbl_prophet}
\alias{residuals.fbl_prophet}
\title{Extract model residuals}
\usage{
\method{residuals}{fbl_prophet}(object, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{...}{Other arguments passed to methods}
}
\value{
A vector of residuals.
}
\description{
Extracts the residuals from an estimated Prophet model.
}
