% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{package}
\name{fable.prophet-package}
\alias{fable.prophet}
\alias{fable.prophet-package}
\title{fable.prophet: Prophet Modelling Interface for 'fable'}
\description{
Allows prophet models from the 'prophet' package to be used in a tidy workflow with the modelling interface of 'fabletools'. This extends 'prophet' to provide enhanced model specification and management, performance evaluation methods, and model combination tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.mitchelloharawild.com/fable.prophet}
  \item Report bugs at \url{https://github.com/mitchelloharawild/fable.prophet/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

Other contributors:
\itemize{
  \item Sean Taylor (Prophet library, https://facebook.github.io/prophet/) [contributor]
  \item Ben Letham (Prophet library, https://facebook.github.io/prophet/) [contributor]
}

}
\keyword{package}
