% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{prophet}
\alias{prophet}
\title{Prophet framework modelling}
\usage{
prophet(...)
}
\arguments{
\item{...}{Arguments used to fit the model}
}
\description{
Prepares a Prophet model for use within the \code{fable} package.
}
\section{Specials}{


\subsection{growth}{
The \code{growth} special is used to specify the trend parameters.
\tabular{ll}{
\code{type}                    \tab The type of seasonality.\cr
\code{capacity}                \tab The carrying capacity for when \code{type} is "logistic".\cr
\code{floor}                   \tab The saturating minimum for when \code{type} is "logistic".\cr
\code{changepoints}            \tab A vector of dates/times for changepoints. If \code{NULL}, changepoints are automatically selected.\cr
\code{n_changepoints}          \tab The total number of changepoints to be selected if \code{changepoints} is \code{NULL}\cr
\code{changepoint_range}       \tab Proportion of the start of the time series where changepoints are automatically selected.\cr
\code{changepoint_prior_scale} \tab Controls the flexibility of the trend.
}
}
}

\examples{

library(tsibble)
tsibbledata::ausretail \%>\%
  filter(Industry == "Cafes, restaurants and catering services") \%>\%
  model(
    prophet = prophet(Turnover ~ season("year", 4, type = "multiplicative"))
  )

}
\seealso{
\itemize{
\item \href{https://facebook.github.io/prophet/}{Prophet homepage}
\item \href{https://cran.r-project.org/web/packages/prophet/index.html}{Prophet R package}
\item \href{https://pypi.org/project/fbprophet/}{Prophet Python package}
}
}
