% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{tidy.fbl_prophet}
\alias{tidy.fbl_prophet}
\title{Extract estimated coefficients from a prophet model}
\usage{
\method{tidy}{fbl_prophet}(x, ...)
}
\arguments{
\item{x}{An object to be converted into a tidy \code{\link[tibble:tibble]{tibble::tibble()}}.}

\item{...}{Additional arguments to tidying method.}
}
\description{
Extract estimated coefficients from a prophet model
}
